+++ frontend/AdoptAPet/src/app/models/user.model.ts
